bfore computer shutodnw
